+++ noark-extraction-validator/src/main/java/com/documaster/validator/main/Run.java
@@ -21,6 +21,7 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
+import com.documaster.validator.config.commands.Noark54Command;
 import org.apache.commons.configuration.ConfigurationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -68,6 +69,7 @@ public class Run {
 
 		Map<String, Command> commands = new HashMap<>();
 		commands.put(Noark53Command.COMMAND_NAME, new Noark53Command(argParser));
+		commands.put(Noark54Command.COMMAND_NAME, new Noark54Command(argParser));
 		for (Command command : commands.values()) {
 			argParser.addCommand(command);
 		}
