+++ noark-extraction-validator/src/main/resources/noark5/noark54/noark54.properties
@@ -1,0 +1,7 @@
+# Noark 5.4 package checksums
+schemaChecksums.arkivstruktur.xsd = f0533d652e66a743f475d01aae94fcb81b0abf3258b4e0adb0d51522702f24e0
+schemaChecksums.metadatakatalog.xsd = 43896558d3882ba44618a57df9fd8e3d0ba714fb26cd5c0aa84d03e7dbdd1431
+schemaChecksums.addml.xsd = 4dc20e5dbf21e94cf211464e9a60b99acd24c45b46ecbf872a32fe2384b904e6
+schemaChecksums.endringslogg.xsd = 6ffed4b158578dcad3a571f9090cf8b728309f28d8210a02e9c5705e435ad5c6
+schemaChecksums.loependeJournal.xsd = 8a05867384eca031cbe2cd3af5d4d3154a27afc7ae18682b9dde7d1f1a0f76d5
+schemaChecksums.offentligJournal.xsd = 55890f63184f4aee1e59ab0cfd782e5df47dc8854805ab2cc5e07dac16a76d9d
